requirements.txt
# --- Core API & Web Server ---
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
websockets>=10.0
python-dotenv>=1.0.0

# --- AI & LangChain Framework ---
langchain>=0.2.0
langchain-core>=0.1.0
langchain-community>=0.2.17
langchain-ollama>=0.1.3

# --- Vector Database & RAG ---
chromadb>=0.4.0
langchain-chroma>=0.1.2
sentence-transformers>=2.2.2

# --- Data Processing & Maths ---
numpy>=1.24.0
pydantic>=2.0.0

# --- Optional Utilities ---
requests>=2.31.0
tqdm>=4.65.0